+++ tests/phpunit/integration/hypeJunction/Dropzone/InputDropzoneViewTest.php
@@ -10,7 +10,7 @@ use Elgg\IntegrationTestCase;
 class InputDropzoneViewTest extends IntegrationTestCase {
 
     public function getPluginID(): string {
-        return 'hypeDropzone';
+        return 'hypedropzone';
     }
 
     public function up() {}